requirements.txt
flask
sqlalchemy
mysql-connector-python
feedparser
flask_wtf
wtforms
flask_login
flask_sqlalchemy
scipy
numpy
pandas
matplotlib
sklearn
alpha_vantage